Cleanup portfolio code
I didn't like the HTML for the portfolio section, and thought it could
be cleaned up a bit with the use of the rails `content_tag` helper. I
also updated the tests (not sure how they were even passing before) to
be more lightweight and test things actually _inside_ the app. An
example of this would be changing an expect statement to look for a link
to Github, as opposed to clicking that link and looking for text on the
github.com page itself.